+++ OrangeDesignSystemDemo/OrangeDesignSystemDemo/Views/Components/Pages/Sliders/SlidersPageVariants.swift
@@ -1,0 +1,8 @@
+//
+//  SlidersPageVariants.swift
+//  OrangeDesignSystemDemo
+//
+//  Created by Jean-Luc CHEVALIER on 25/11/2022.
+//
+
+import Foundation